requirements.txt
google-generativeai
supabase
python-dotenv
Pillow
requests
httpx